// src/src/features/Home/types.ts
export interface ExchangeRateItem {
    title: string;
    name: string;
    price: string;
    rateOfChange: string;
    category: string;
    highestRate: string;
    lowestRate: string;
    sourceCreated: string | Date;
}